server/requirements.txt
openenv-core>=0.2.2
fastapi>=0.104.0
uvicorn>=0.24.0
fastmcp>=3.0.0
pydantic>=2.0.0
python-multipart>=0.0.6